+++ packages/react/src/components/UserInformation/UserInformation.js
@@ -8,6 +8,7 @@ import { Avatar } from '../Avatar';
 import { Icon } from '../Icon';
 import { Throbber } from '../Throbber';
 import { appendClassNames } from '../../lib/appendClassNames';
+import formatTimestamp from '../../lib/formatTimestamp';
 
 const userSidebarCss = css`
   padding: 0 1rem 1rem;
@@ -46,25 +47,6 @@ const UserInformation = () => {
     const host = RCInstance.getHost();
     const URL = `${host}/avatar/${username}`;
     return URL;
-  };
-
-  const formatTimestamp = (timestamp) => {
-    const date = new Date(timestamp);
-    const now = new Date();
-
-    const isDifferentDay =
-      date.getDate() !== now.getDate() ||
-      date.getMonth() !== now.getMonth() ||
-      date.getFullYear() !== now.getFullYear();
-
-    const options = { hour: 'numeric', minute: 'numeric', hour12: true };
-    const formattedTime = date.toLocaleTimeString('en-US', options);
-
-    return isDifferentDay
-      ? `${date.toLocaleDateString('en-US', {
-          weekday: 'long',
-        })} ${formattedTime}`
-      : formattedTime;
   };
 
   useEffect(() => {
